refactor(friends): split friends model
