+++ src/app/main/sidebar/contact/mod.rs
@@ -1,72 +1,29 @@
-mod friends_group;
+mod friends;
 mod group_item;
 
 use relm4::factory::FactoryVecDeque;
-use relm4::{adw, gtk, ComponentParts, ComponentSender, SimpleComponent, WidgetPlus};
+use relm4::{
+    adw, gtk, Component, ComponentController, ComponentParts, ComponentSender, Controller,
+    SimpleComponent, WidgetPlus,
+};
 use std::cell::RefCell;
 
 use adw::{prelude::*, ViewStack, ViewSwitcherBar};
 use gtk::{Box, Button, Entry, EntryIconPosition, ListBox, Orientation, ScrolledWindow};
 use tokio::task;
 
+use friends::FriendsModel;
+use crate::db::sql::{get_db, refresh_groups_list, Group};
 use super::SidebarMsg;
-use crate::db::sql::{get_db, refresh_friends_list, refresh_groups_list, Friend, Group};
-use friends_group::FriendsGroup;
 
 #[derive(Debug)]
 pub struct ContactModel {
-    friends_list: Option<RefCell<FactoryVecDeque<Box, FriendsGroup, ContactMsg>>>,
+    friends: Controller<FriendsModel>,
     groups_list: Option<RefCell<FactoryVecDeque<ListBox, Group, ContactMsg>>>,
-    is_refresh_friends_button_enabled: bool,
     is_refresh_groups_button_enabled: bool,
 }
 
 impl ContactModel {
-    fn render_friends(&self) -> rusqlite::Result<()> {
-        let mut friends_list = self.friends_list.as_ref().unwrap().borrow_mut();
-        friends_list.clear();
-
-        let conn = get_db();
-
-        let mut stmt = conn.prepare("Select id, name, remark, group_id from friends")?;
-        let friends: Vec<Friend> = stmt
-            .query_map([], |row| {
-                Ok(Friend {
-                    id: row.get(0)?,
-                    name: row.get(1)?,
-                    remark: row.get(2)?,
-                    group_id: row.get(3)?,
-                })
-            })?
-            .map(|result| result.unwrap())
-            .collect();
-
-        let friends_groups: Vec<FriendsGroup> = conn
-            .prepare("Select id, name, online_friends from friends_groups")?
-            .query_map([], |row| {
-                Ok(FriendsGroup {
-                    id: row.get(0)?,
-                    name: row.get(1)?,
-                    online_friends: row.get(2)?,
-                    friends: friends
-                        .clone()
-                        .into_iter()
-                        .filter(|friend| friend.group_id == row.get(0).unwrap())
-                        .collect(),
-                })
-            })?
-            .map(|result| result.unwrap())
-            .collect();
-
-        for friends_group in friends_groups {
-            friends_list.push_back(friends_group);
-        }
-
-        friends_list.render_changes();
-
-        Ok(())
-    }
-
     fn render_groups(&self) -> rusqlite::Result<()> {
         let mut groups_list = self.groups_list.as_ref().unwrap().borrow_mut();
         groups_list.clear();
@@ -93,16 +50,6 @@ impl ContactModel {
     }
 }
 
-async fn refresh_friends(sender: ComponentSender<ContactModel>) {
-    sender.output(SidebarMsg::PushToast(
-        "Start refreshing the friends list...".to_string(),
-    ));
-    match refresh_friends_list().await {
-        Ok(_) => sender.input(ContactMsg::RenderFriends),
-        Err(err) => sender.output(SidebarMsg::PushToast(err.to_string())),
-    }
-}
-
 async fn refresh_groups(sender: ComponentSender<ContactModel>) {
     sender.output(SidebarMsg::PushToast(
         "Start refreshing the groups list...".to_string(),
@@ -116,17 +63,16 @@ async fn refresh_groups(sender: ComponentSender<ContactModel>) {
 #[derive(Debug)]
 pub enum ContactMsg {
     SelectChatroom(i64, bool),
-    RefreshFriends,
     RefreshGroups,
-    RenderFriends,
     RenderGroups,
+    PushToast(String),
 }
 
 #[relm4::component(pub)]
 impl SimpleComponent for ContactModel {
     type Input = ContactMsg;
     type Output = SidebarMsg;
-    type Widgets = MainPageWidgets;
+    type Widgets = ContactWIdgets;
     type InitParams = ();
 
     view! {
@@ -140,34 +86,6 @@ impl SimpleComponent for ContactModel {
             ViewSwitcherBar {
                 set_stack: Some(&contact_stack),
                 set_reveal: true
-            }
-        },
-        contact_friends = Box {
-            set_orientation: Orientation::Vertical,
-            Box {
-                set_margin_all: 8,
-                Button {
-                    #[watch]
-                    set_sensitive: model.is_refresh_friends_button_enabled,
-                    set_tooltip_text: Some("Refresh friends list"),
-                    set_icon_name: "view-refresh-symbolic",
-                    set_margin_end: 8,
-                    connect_clicked[sender] => move |_| {
-                        sender.input(ContactMsg::RefreshFriends);
-                    },
-                },
-                #[name = "search_friends_entry"]
-                Entry {
-                    set_icon_from_icon_name: (EntryIconPosition::Secondary, Some("system-search-symbolic")),
-                    set_placeholder_text: Some("Search in friends..."),
-                    set_width_request: 320 - 3 * 8 - 32
-                },
-            },
-            ScrolledWindow {
-                set_child: contact_friends_list = Some(&Box) {
-                    set_vexpand: true,
-                    set_orientation: Orientation::Vertical,
-                }
             }
         },
         contact_groups = Box {
@@ -214,30 +132,27 @@ impl SimpleComponent for ContactModel {
         sender: &ComponentSender<Self>,
     ) -> ComponentParts<Self> {
         let mut model = ContactModel {
-            friends_list: None,
+            friends: FriendsModel::builder()
+                .launch(())
+                .forward(&sender.input, |message| message),
             groups_list: None,
-            is_refresh_friends_button_enabled: true,
             is_refresh_groups_button_enabled: true,
         };
         let widgets = view_output!();
 
         let contact_stack: &ViewStack = &widgets.contact_stack;
 
-        let friends = contact_stack.add_titled(&widgets.contact_friends, None, "Friends");
+        let friends = contact_stack.add_titled(model.friends.widget(), None, "Friends");
         let groups = contact_stack.add_titled(&widgets.contact_groups, None, "Groups");
 
         friends.set_icon_name(Some("person2-symbolic"));
         groups.set_icon_name(Some("people-symbolic"));
 
-        let friends_list: FactoryVecDeque<Box, FriendsGroup, ContactMsg> =
-            FactoryVecDeque::new(widgets.contact_friends_list.clone(), &sender.input);
         let groups_list: FactoryVecDeque<ListBox, Group, ContactMsg> =
             FactoryVecDeque::new(widgets.contact_groups_list.clone(), &sender.input);
 
-        model.friends_list = Some(RefCell::new(friends_list));
         model.groups_list = Some(RefCell::new(groups_list));
 
-        model.render_friends().unwrap();
         model.render_groups().unwrap();
 
         ComponentParts { model, widgets }
@@ -249,22 +164,9 @@ impl SimpleComponent for ContactModel {
             SelectChatroom(account, is_group) => {
                 sender.output(SidebarMsg::SelectChatroom(account, is_group));
             }
-            RefreshFriends => {
-                self.is_refresh_friends_button_enabled = false;
-                task::spawn(refresh_friends(sender.clone()));
-            }
             RefreshGroups => {
                 self.is_refresh_groups_button_enabled = false;
                 task::spawn(refresh_groups(sender.clone()));
-            }
-            RenderFriends => {
-                match self.render_friends() {
-                    Ok(_) => sender.output(SidebarMsg::PushToast(
-                        "Refreshed the friends list.".to_string(),
-                    )),
-                    Err(err) => sender.output(SidebarMsg::PushToast(err.to_string())),
-                }
-                self.is_refresh_friends_button_enabled = true;
             }
             RenderGroups => {
                 match self.render_groups() {
@@ -274,6 +176,9 @@ impl SimpleComponent for ContactModel {
                     Err(err) => sender.output(SidebarMsg::PushToast(err.to_string())),
                 }
                 self.is_refresh_groups_button_enabled = true;
+            }
+            PushToast(msg) => {
+                sender.output(SidebarMsg::PushToast(msg));
             }
         }
     }
